requirements.txt
# 依赖库列表
spire.presentation
requests
websocket-client
Pillow